requirements.txt
pyyaml
pyperclip